Project 3
The stressed vowel sound (i.e., the primary stress) in both words must be identical, as well as any subsequent sounds. For example, "sky" and "high"; "skylight" and "highlight".
The sound phoneme that precedes the stressed vowel in the words must not be the same. For example, "bean" and "green" is a perfect rhyme, while "leave" and "believe" is not.
In the rest of this document, "rhyme" refers to perfect rhymes unless otherwise specified.

Output:
Print the words in upper case, sorted alphabetically, one word per line.
If no rhyming word is found for the user-provided word, does not print anything.
Assumes that the user-provided word will always exist in the pronunciation dictionary and will have a primary stress phoneme.
Does not remove duplicates.