Project 9
A) Huffman
One way to create a prefix code is called Huffman Coding. It generates binary encodings, which map nicely to binary trees if we allow the left child references to represent 0 and the right to represent 1. With a given non-hierarchical representation of the tree and an encoded sequence of values, the project is centered to reconstruct the tree from the given representation, and use it to decode a given sequence of values.

B) Writer Bot
A method called Markov Chain Analysis determines the probability that certain words will follow combinations of other words in a text. This program uses Markov chain analysis statistics to generate new text ensuring that the new text will have similar statistical properties to the original.